+++ src/enum.ts
@@ -1,4 +1,4 @@
 export const ADD_COLOR = 'themeChanger.addColor';
-export const CLEAR_CLORS = 'themeChanger.clearColors';
+export const CLEAR_COLORS = 'themeChanger.clearColors';
 export const CHANGE_THEME_COLOR = 'themeChanger.changeThemeColor';
 export const COLORS_VIEW = 'themeChanger.sidebar-colorsView';